프로필 페이지, 프로필 수정, 로그인 할때만 할 수 있는 기능 구현

+++ Python/Django/Practice/20221017_19/article/views.py
@@ -1,14 +1,16 @@
 from django.shortcuts import render, redirect
 from .forms import ReviewForm, CommentForm
 from .models import Review, Comment
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 def index(request):
   context = {
-    'reviews' : Review.objects.all(),
+    'reviews' : Review.objects.all().order_by('-updated_at'),
   }
   return render(request, 'article/index.html', context)
 
+@login_required
 def create(request):
   if request.method == 'POST':
     review_form = ReviewForm(request.POST, request.FILES)
@@ -25,6 +27,7 @@ def create(request):
   }
   return render(request, 'article/create.html', context)
 
+@login_required
 def detail(request, pk):
   review = Review.objects.get(pk=pk)
   # 댓글을 써야할 모델폼을 가지고 온다
@@ -44,6 +47,7 @@ def detail(request, pk):
   }
   return render(request, 'article/detail.html', context)
 
+@login_required
 def comment_create(request, pk):
   review = Review.objects.get(pk=pk)
   if request.method == 'POST':
@@ -55,27 +59,42 @@ def comment_create(request, pk):
       comment.save()
   return redirect('article:detail', review.pk)
 
+@login_required
 def comment_delete(request, pk, detail_pk):
   detail = Review.objects.get(pk=detail_pk)
   comment = Comment.objects.get(pk=pk)
-  comment.delete()
-  return redirect('article:detail', detail.pk)
 
+  if request.user == comment.user:
+    comment.delete()
+    return redirect('article:detail', detail.pk)
+  else:
+    return redirect('article:detail', detail.pk)
+
+@login_required
 def update(request, pk):
   review = Review.objects.get(pk=pk)
-  if request.method == 'POST':
-    update_form = ReviewForm(request.POST, request.FILES, instance=review)
-    if update_form.is_valid():
-      update_form.save()
-      return redirect('article:detail', review.pk)
-  else:
-    update_form = ReviewForm(instance=review)
-  context = {
-    'update_form' : update_form,
-    'num' : review.pk,
-  }
-  return render(request, 'article/update.html', context)
 
+  if review.user == request.user:
+    if request.method == 'POST':
+      update_form = ReviewForm(request.POST, request.FILES, instance=review)
+      if update_form.is_valid():
+        update_form.save()
+        return redirect('article:detail', review.pk)
+    else:
+      update_form = ReviewForm(instance=review)
+    context = {
+      'update_form' : update_form,
+      'num' : review.pk,
+    }
+    return render(request, 'article/update.html', context)
+  else:
+    return redirect('article:detail', review.pk)
+
+@login_required
 def delete(request, pk):
-  Review.objects.get(pk=pk).delete()
-  return redirect('article:index')
+  review = Review.objects.get(pk=pk)
+  if request.user == review.user:
+    review.delete()
+    return redirect('article:index')
+  else:
+    return redirect('article:detail', review.pk)
